effects module is added
